[WIP] refactor test utils
